refactor: main css - reducing duplication and add mixins
