+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Nav } from "./components/Nav/Nav";
-import { Table } from "./components/Main/Table";
+import { Table } from "./components/Main/Table/Table";
 import { Header } from "./components/Main/Header/Header";
 
 function App() {
