--- src/components/Main/index.module.scss
@@ -1,116 +1,0 @@
-.table {
-  display: grid;
-  border-collapse: collapse;
-  min-width: 100%;
-  grid-template-columns:
-          minmax(60px, 0.7fr)
-    minmax(100px, 1.1fr)
-    minmax(100px, 1.27fr)
-    minmax(100px, 1.17fr)
-    minmax(100px, 1.33fr)
-    minmax(100px, 2.67fr)
-    minmax(100px, 1.33fr)
-    minmax(20px, 0.5fr);
-  font-size: 0.8rem;
-  row-gap: 4px;
-  padding-top: 50px;
-}
-
-.list {
-  background-color: var(--contrast-gray);
-}
-
-.header,
-.cards,
-.list {
-  display: contents;
-}
-
-.headItem,
-.item {
-  padding: 15px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.headItem {
-  position: sticky;
-  top: 0;
-  text-align: left;
-  font-weight: normal;
-  font-size: 1.1rem;
-  color: white;
-}
-
-.headItem:last-child {
-  border: 0;
-}
-
-.item {
-  padding-top: 10px;
-  padding-bottom: 10px;
-  color: #808080;
-}
-
-.headItem {
-  font-size: 0.7rem;
-}
-
-.headItem, .item {
-  padding: 5px;
-  display: grid;
-  align-content: center;
-}
-
-.list > .item {
-  color: var(--white);
-
-  &.orange span {
-    color: var(--orange-text);
-    background-color: var(--orange-bg);
-    width: 93px;
-    height: 18px;
-    font-size: 0.6rem;
-    border-radius: var(--basic-border);
-    display: grid;
-    justify-content: center;
-    align-content: center;
-  }
-
-  &.green span {
-    color: var(--green-text);
-    background-color: var(--green-bg);
-    width: 50px;
-    height: 18px;
-    font-size: 0.6rem;
-    border-radius: var(--basic-border);
-    display: grid;
-    justify-content: center;
-    align-content: center;
-  }
-
-  &:first-child {
-    border-top-left-radius: var(--basic-border);
-    border-bottom-left-radius: var(--basic-border);
-  }
-
-  &:last-child {
-    border-top-right-radius: var(--basic-border);
-    border-bottom-right-radius: var(--basic-border);
-  }
-}
-
-.list > ul > .item:last-child {
-  color: var(--text-gray);
-  font-size: 0.65rem;
-}
-
-.item {
-  color: var(--white);
-  background-color: var(--contrast-gray);
-}
-
-.table .list:hover .item {
-  background-color: var(--hover-items);
-}
