+++ src/components/Nav/index.module.scss
@@ -12,7 +12,7 @@
 
 .btn {
   height: 36px;
-  font-size: 0.8rem;
+  font-size: 0.85rem;
   padding: 8px;
   border-radius: var(--basic-border);
