+++ src/features/Select/index.module.scss
@@ -1,3 +1,5 @@
+@import "../../styles/mixins";
+
 .select {
   position: relative;
   background: var(--gray);
@@ -7,6 +9,11 @@
   flex-basis: 100%;
   flex-grow: 1;
   cursor: pointer;
+
+  span {
+    @include ellipsis;
+  }
+
   &.active {
     background-color: var(--whitest-gray);
   }
@@ -17,12 +24,12 @@
   gap: 10px;
   width: 100%;
   height: 100%;
-  font-size: 0.8rem;
+  font-size: 0.85rem;
   align-items: center;
   justify-content: space-between;
   padding: 8px;
 }
 
 .arrow {
-  width: 15px;
+  width: 20px;
 }
